DGV puede seleccionar, modificar, calcular total linea + inicio de crear articulos de pedido (interrumpido) + modficacion de nombre de roles en Backend

+++ MTACCore/Entidades/PedidoArticulos.cs
@@ -10,9 +10,9 @@ namespace MTCore_AC.Entidades
         public string id_pedido { get; set; }
         public int id_articulo { get; set; }
         public int cantidad { get; set; }
-        public float precio_unidad { get; set; }
+        public decimal precio_unidad { get; set; }
         public PedidoArticulos() {}
-        public PedidoArticulos(string IdPedido, int IdArticulo, int Cantidad, float PrecioUnidad) {
+        public PedidoArticulos(string IdPedido, int IdArticulo, int Cantidad, decimal PrecioUnidad) {
             this.id_pedido = IdPedido;
             this.id_articulo = IdArticulo;   
             this.cantidad = Cantidad;
